+++ src/main/java/pl/zycienakodach/pragmaticflights/modules/pricing/api/events/DiscountApplied.java
@@ -1,4 +1,6 @@
 package pl.zycienakodach.pragmaticflights.modules.pricing.api.events;
 
-public record DiscountApplied(String orderId, double discountInEuro) implements PricingEvent {
+import java.math.BigDecimal;
+
+public record DiscountApplied(String orderId, BigDecimal discountInEuro) implements PricingEvent {
 }
